+++ eanalytics_api_py/__init__.py
@@ -1,1 +1,1 @@
-from . jupyter import jupyter_helper
+from .conn import Conn